work in progress
Former-commit-id: 203eb11c6198223253353151840e0ed9c548bc08
Former-commit-id: 3bc921fd62d528262765f6a5f652e53b9f69b937

+++ src/socket_util.h
@@ -183,17 +183,17 @@ typedef struct {
     /*
      * THING_STATE_BIT_SHIFT_EXT1_WEAPON_ID_PRESENT
      */
-    uint8_t weapon_id;
+    uint16_t weapon_id;
 
     /*
      * THING_STATE_BIT_SHIFT_EXT1_SHIELD_ID_PRESENT
      */
-    uint8_t shield_id;
+    uint16_t shield_id;
 
     /*
      * THING_STATE_BIT_SHIFT_EXT2_MAGIC_ID_PRESENT
      */
-    uint8_t magic_id;
+    uint16_t magic_id;
 
     /*
      * THING_STATE_BIT_SHIFT_EXT1_EFFECT_PRESENT
